Refactor: Move AI types and remove unused test files
