+++ _archive/temp_hive/app/api/cortex/logs/route.ts
