+++ _archive/temp_hive/components/HiveMind.tsx
