+++ _archive/temp_hive/components/SensoryCortex.tsx
