+++ _archive/temp_hive/components/StrataShell.tsx
