+++ _archive/temp_hive/cortex/observer.ts
