+++ _archive/temp_hive/docs/genesis-records/2025-11-19T20-48-29Z-first-collaboration.md
