+++ _archive/temp_hive/docs/genesis-records/genesis-2025-11-19-codex.md
