+++ shared/types/ai.ts
